allow instance name to be set, use apache httpclient
