+++ build.gradle
@@ -28,6 +28,8 @@ dependencies {
     compile 'org.springframework.boot:spring-boot-starter-actuator:2.0.6.RELEASE'
 
     compile 'org.springframework.cloud:spring-cloud-starter-openfeign:2.0.2.RELEASE'
+    compile 'io.github.openfeign:feign-httpclient:9.7.0'
+    compile 'org.apache.httpcomponents:httpclient:4.5.6'
 
     compile 'com.microsoft.azure:azure-keyvault:1.0.0'
     compile 'com.microsoft.azure:adal4j:1.5.0'
@@ -35,7 +37,7 @@ dependencies {
     compile 'org.springframework.ldap:spring-ldap-core:2.3.2.RELEASE'
     compile 'com.microsoft.azure:applicationinsights-core:2.2.0'
     compile 'com.microsoft.azure:applicationinsights-web:2.2.0'
-    compile 'com.microsoft.azure:applicationinsights-spring-boot-starter:1.0.1-BETA'
+    compile 'com.microsoft.azure:applicationinsights-spring-boot-starter:1.1.0-BETA'
     compile 'com.microsoft.azure:applicationinsights-logging-logback:2.2.0'
 
     compile 'org.projectlombok:lombok:1.18.2'
