--- src/main/java/uk/gov/hmcts/reform/idam/health/ScheduleConfig.java
@@ -1,20 +1,0 @@
-package uk.gov.hmcts.reform.idam.health;
-
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.TaskScheduler;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-
-@Configuration
-public class ScheduleConfig {
-
-    @Bean
-    public TaskScheduler taskScheduler(){
-        ThreadPoolTaskScheduler threadPoolTaskScheduler = new ThreadPoolTaskScheduler();
-        threadPoolTaskScheduler.setPoolSize(5);
-        threadPoolTaskScheduler.setThreadNamePrefix("ProbeScheduler");
-        return threadPoolTaskScheduler;
-    }
-
-}
